requirements.txt
revtok==0.0.3
tqdm==4.27
torchvision==0.2.1
mo-future>=2.20
pyparsing==2.3.0
wandb==0.8.30
n2w
matplotlib==3.0.2
transformers==2.8.0
asciitree
apex
nltk==3.4.5
rapidfuzz==0.9.1

# wikisql
records
babel
tabulate
scipy

# demo use only
sqlalchemy==1.3.20

jonlines

#sql parser
mo_sql_parsing

# neo4j graph
py2neo
neo4j

# neural network
pytorch
transformers